seperation of ui and sdk
